+++ __init__.py
@@ -2,4 +2,4 @@
 # Audio-to-SRT Toolkit package initialization
 # 2025-04-23 -JS
 
-__version__ = '0.0.020'  # Follow semantic versioning (major.minor.patch) - 2025-04-23 -JS
+__version__ = '0.0.021'  # Follow semantic versioning (major.minor.patch) - 2025-04-23 -JS